Refactor out common non-NIP-01 tags out of NostrEvent file into separate protocols for better code separation
